Create globalmap_server_nodelet.hpp
